requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain_huggingface
beautifulsoup4